move beam related class from shared to 3d_buliding
