+++ src/for_3D_build/particle_dynamics/solid_dynamics/slender_structure_dynamics.h
@@ -37,6 +37,7 @@
 #include "solid_particles.h"
 #include "elastic_solid.h"
 #include "slender_structure_math.h"
+#include "beam_particles.h"
 
 namespace SPH
 {
