+++ src/for_3D_build/particle_dynamics/solid_dynamics/slender_structure_math.h
@@ -34,7 +34,7 @@
 #include "base_kernel.h"
 #include "elastic_solid.h"
 #include "weakly_compressible_fluid.h"
-
+#include "beam_particles.h"
 namespace SPH
 {
 namespace slender_structure_dynamics
@@ -44,24 +44,24 @@ namespace slender_structure_dynamics
  * when the axis about which they occur points toward the observer,
  * and the coordinate system is right-handed.
  */
-Vec2d getVectorAfterThinStructureRotation(const Vec2d &initial_vector, const Vec2d &rotation_angles);
+//Vec2d getVectorAfterThinStructureRotation(const Vec2d &initial_vector, const Vec2d &rotation_angles);
 Vec3d getVectorAfterThinStructureRotation(const Vec3d &initial_vector, const Vec3d &rotation_angles);
 
 /** Vector change rate after rotation. */
-Vec2d getVectorChangeRateAfterThinStructureRotation(const Vec2d &initial_vector, const Vec2d &rotation_angles, const Vec2d &angular_vel);
+//Vec2d getVectorChangeRateAfterThinStructureRotation(const Vec2d &initial_vector, const Vec2d &rotation_angles, const Vec2d &angular_vel);
 Vec3d getVectorChangeRateAfterThinStructureRotation(const Vec3d &initial_vector, const Vec3d &rotation_angles, const Vec3d &angular_vel);
 
 /** get the rotation from pseudo-normal for finite deformation. */
-Vec2d getRotationFromPseudoNormalForFiniteDeformation(const Vec2d &dpseudo_n_d2t, const Vec2d &rotation, const Vec2d &angular_vel, Real dt);
+//Vec2d getRotationFromPseudoNormalForFiniteDeformation(const Vec2d &dpseudo_n_d2t, const Vec2d &rotation, const Vec2d &angular_vel, Real dt);
 Vec3d getRotationFromPseudoNormalForFiniteDeformation(const Vec3d &dpseudo_n_d2t, const Vec3d &rotation, const Vec3d &angular_vel, Real dt);
 
 /** get the rotation from pseudo-normal for small deformation. */
-Vec2d getRotationFromPseudoNormalForSmallDeformation(const Vec2d &dpseudo_n_d2t, const Vec2d &rotation, const Vec2d &angular_vel, Real dt);
+//Vec2d getRotationFromPseudoNormalForSmallDeformation(const Vec2d &dpseudo_n_d2t, const Vec2d &rotation, const Vec2d &angular_vel, Real dt);
 Vec3d getRotationFromPseudoNormalForSmallDeformation(const Vec3d &dpseudo_b_n_d2t, const Vec3d &dpseudo_n_d2t, const Vec3d &rotation, const Vec3d &angular_vel, Real dt);
 Vec3d getRotationFromPseudoNormalForSmallDeformation_b(const Vec3d &dpseudo_b_n_d2t, const Vec3d &dpseudo_n_d2t, const Vec3d &rotation, const Vec3d &angular_vel, Real dt);
 
 /** get the current normal direction from deformation gradient tensor. */
-Vec2d getNormalFromDeformationGradientTensor(const Mat2d &F);
+//Vec2d getNormalFromDeformationGradientTensor(const Mat2d &F);
 Vec3d getNormalFromDeformationGradientTensor(const Mat3d &F);
 Vec3d getBinormalFromDeformationGradientTensor(const Mat3d &F);
 
@@ -79,15 +79,15 @@ Vecd getWENORightState(const Vecd &e_ij, const Real &r_ij, const Vecd &particle_
                        const Matd &gradient_particle_i_value, const Vecd &particle_j_value, const Matd &gradient_particle_j_value);
 
 /** get the artificial rotation from the pseudo-normal jump. */
-Vec2d getRotationJump(const Vec2d &pseudo_n_jump, const Mat2d &transformation_matrix);
+//Vec2d getRotationJump(const Vec2d &pseudo_n_jump, const Mat2d &transformation_matrix);
 Vec3d getRotationJump(const Vec3d &pseudo_n_jump, const Mat3d &transformation_matrix);
 
 /** get the corrected Eulerian Almansi strain tensor according to plane stress problem. */
-Mat2d getCorrectedAlmansiStrain(const Mat2d &current_local_almansi_strain, const Real &nu_);
+//Mat2d getCorrectedAlmansiStrain(const Mat2d &current_local_almansi_strain, const Real &nu_);
 Mat3d getCorrectedAlmansiStrain(const Mat3d &current_local_almansi_strain, const Real &nu_);
 
 /** get the correction matrix. */
-Mat2d getCorrectionMatrix(const Mat2d &local_deformation_part_one);
+//Mat2d getCorrectionMatrix(const Mat2d &local_deformation_part_one);
 Mat3d getCorrectionMatrix(const Mat3d &local_deformation_part_one);
 Mat3d getCorrectionMatrix_beam(const Mat3d &local_deformation_part_one);
 } // namespace slender_structure_dynamics
