+++ src/shared/particle_generator/base_particle_generator.cpp
@@ -48,21 +48,7 @@ void SurfaceParticleGenerator::initializeSurfaceProperties(const Vecd &surface_n
 }
 //=================================================================================================//
 
-LineParticleGenerator::LineParticleGenerator(SPHBody &sph_body)
-    : ParticleGenerator(sph_body),
-      n_(*base_particles_.getVariableByName<Vecd>("NormalDirection")),
-      thickness_(*base_particles_.getVariableByName<Real>("Thickness")),
-      b_n_(*base_particles_.getVariableByName<Vecd>("BinormalDirection")),
-      width_(*base_particles_.getVariableByName<Real>("Width")) {}
-//=================================================================================================//
-void LineParticleGenerator::initializeLineProperties(const Vecd &line_normal, const Vecd &line_binormal, Real thickness, Real width)
-{
-    n_.push_back(line_normal);
-    thickness_.push_back(thickness);
-    b_n_.push_back(line_binormal);
-    width_.push_back(width);
-}
-//=================================================================================================//
+
 void ObserverParticleGenerator::initializeGeometricVariables()
 {
     for (size_t i = 0; i < positions_.size(); ++i)
