+++ tests/3d_examples/test_3d_slender_beam/test_3d_slender_beam.cpp
@@ -9,6 +9,8 @@
 #include <gtest/gtest.h>
 #include "slender_structure_dynamics.h"
 #include "slender_structure_math.h"
+#include "beam_particles.h"
+#include "beam_particle_generator.h"
 using namespace SPH;
 
 /**
